test(observer): mirror the spike package with per-module test files (#2348)
The test_mirrors_source hard convention expects a package to be mirrored by
tests/test_<pkg>_*.py or tests/test_<pkg>/; a single tests/test_<pkg>.py
matches neither, so splitting the POC into a package left it unmirrored.

Split the suite along the same seam as the source: _ranking.py covers the pure
math (attribution, cohort excess, control availability, weakest signal, the
wc -l ship gate, thresholds) and _report.py covers substrate access, the two
read-model helpers, and rendering. Shared row/coverage/record builders move to
structural_decay_test_helpers.py, following the existing *_test_helpers.py
convention, so neither test module imports the other.

Adds three tests the split made natural: rendered entries carry their evidence
and exactly one weakest-signal line each, panel size is reported as derived
from the decoded record, and coverage counts measured runs that do not join.

Verified: no hard convention violation mentions structural_decay, and
net-new blocking violations against origin/main is 0.
